[AIRToAIE] Address PR review: guard refeed_count range, robust test matches
Copilot review feedback:

- Range/truncation (getLockForDMA wlockInit, allocateCoreLocksPerMemcpyOp
  setInit): air.refeed_count is an int64 attribute narrowed into 32-bit AIE
  semaphore lock init/acquire values; a negative or >INT32_MAX value would
  silently truncate and mis-synchronize/deadlock. Guard at the single reader
  air::getRefeedCount: values <= 1 -> 1 (no re-feed), values > INT32_MAX -> warn
  and fall back to 1. Both narrowing call sites are now provably safe; annotate
  them. Also tighten air.channel's verifier to reject a refeed_count outside
  [1, INT32_MAX] (was only >= 1), catching the malformed channel carrier early,
  and add an overflow verifier test.
- Brittle FileCheck: the memtile lowering test required air.refeed_count to be
  the first attribute printed on aie.buffer. Match it position-independently
  (aie.buffer{{.*}}air.refeed_count) so attribute reordering can't cause false
  negatives.

check-air-mlir: 423 passed, 7 unsupported, 7 xfail.

+++ mlir/lib/Conversion/AIRToAIESchedulingUtils.cpp
@@ -874,11 +874,14 @@ air::DMAAllocator::getLockForDMA(air::MemcpyInterface &memcpyOp,
   // therefore INIT to N -- with the default slot-count init the first
   // AcquireGreaterEqual N can never fire and the producer feeding the buffer
   // deadlocks.
+  // getRefeedCount guarantees 1 <= count <= INT32_MAX, so this init fits the
+  // 32-bit lock without truncation.
   int64_t wlockInit = init;
   if (UsesSemaphoreLocks)
     wlockInit = std::max(wlockInit, air::getRefeedCount(bufferOp));
-  auto wlock =
-      UsesSemaphoreLocks ? allocateLockOp(device, tile, wlockInit) : rlock;
+  auto wlock = UsesSemaphoreLocks
+                   ? allocateLockOp(device, tile, static_cast<int>(wlockInit))
+                   : rlock;
   lock_allocation_list.push_back({bufferOp, air_chan, channel, rlock, wlock});
   return std::make_pair(rlock, wlock);
 }
